Relocate code
