+++ internal/cluster/distribution/eks/eksprovider/workflow/bootstrap_activity.go
@@ -32,7 +32,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	storageUtil "k8s.io/kubernetes/pkg/apis/storage/util"
 
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution/infrastructure/aws/awsworkflow"
+	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 	"github.com/banzaicloud/pipeline/internal/providers/amazon"
 	"github.com/banzaicloud/pipeline/pkg/cadence"
 	"github.com/banzaicloud/pipeline/pkg/k8sclient"
