+++ internal/cluster/distribution/eks/eksprovider/workflow/create_cluster_user_access_key_activity.go
@@ -23,7 +23,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/iam"
 	"go.uber.org/cadence/activity"
 
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution/infrastructure/aws/awsworkflow"
+	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 	"github.com/banzaicloud/pipeline/internal/secret/secrettype"
 	"github.com/banzaicloud/pipeline/pkg/cluster"
 	"github.com/banzaicloud/pipeline/pkg/providers/amazon"
