+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_control_plane_activity.go
@@ -25,7 +25,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/eks"
 	"go.uber.org/cadence/activity"
 
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution/infrastructure/aws/awsworkflow"
+	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 )
 
 const DeleteControlPlaneActivityName = "eks-delete-control-plane"
