+++ internal/cluster/distribution/eks/eksprovider/workflow/delete_ssh_key_activity.go
@@ -24,7 +24,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/iam"
 	"go.uber.org/cadence/activity"
 
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution/infrastructure/aws/awsworkflow"
+	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 )
 
 const DeleteSshKeyActivityName = "eks-delete-ssh-key"
