+++ internal/cluster/distribution/eks/eksprovider/workflow/save_k8s_config_acitivity.go
@@ -26,7 +26,7 @@ import (
 	"go.uber.org/cadence/activity"
 	"go.uber.org/zap"
 
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution/infrastructure/aws/awsworkflow"
+	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 	"github.com/banzaicloud/pipeline/internal/secret/secrettype"
 	"github.com/banzaicloud/pipeline/src/secret"
 	"github.com/banzaicloud/pipeline/src/utils"
