+++ internal/cluster/distribution/eks/eksprovider/workflow/wait_elbs_deletion_activity.go
@@ -24,7 +24,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/elb"
 	"go.uber.org/cadence/activity"
 
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution/infrastructure/aws/awsworkflow"
+	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 )
 
 const WaitELBsDeletionActivityName = "eks-wait-elbs-deletion"
