+++ internal/cluster/distribution/eks/eksworkflow/activity_wait_update_version.go
@@ -24,7 +24,7 @@ import (
 	"go.uber.org/cadence/activity"
 
 	"github.com/banzaicloud/pipeline/internal/cluster/distribution/eks/eksprovider/workflow"
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution/infrastructure/aws/awsworkflow"
+	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 )
 
 const WaitUpdateClusterVersionActivityName = "eks-wait-update-version"
